+++ comms_champion/app/cc_view/src/widget/field/ShortIntValueFieldWidget.cpp
@@ -61,7 +61,7 @@ void ShortIntValueFieldWidget::refreshImpl()
     assert(m_ui.m_serValueLineEdit != nullptr);
     updateValue(*m_ui.m_serValueLineEdit, m_wrapper->getSerialisedString());
 
-    auto value = adjustRealToDisplayed(m_wrapper->getValue());
+    auto value = adjustRealToDisplayed(static_cast<int>(m_wrapper->getValue()));
     assert(m_ui.m_valueSpinBox);
     if (m_ui.m_valueSpinBox->value() != value) {
         m_ui.m_valueSpinBox->setValue(value);
@@ -113,7 +113,7 @@ void ShortIntValueFieldWidget::serialisedValueUpdated(const QString& value)
 
 void ShortIntValueFieldWidget::valueUpdated(int value)
 {
-    if (value == adjustRealToDisplayed(m_wrapper->getValue())) {
+    if (value == adjustRealToDisplayed(static_cast<int>(m_wrapper->getValue()))) {
         return;
     }
